+++ src/world/Objects/DynamicObject.cpp
@@ -92,7 +92,7 @@ void DynamicObject::Create(Unit* caster, Spell* pSpell, float x, float y, float 
     SetEntry(m_spellProto->getId());
     setFloatValue(OBJECT_FIELD_SCALE_X, 1);
     setUInt64Value(DYNAMICOBJECT_CASTER, caster->GetGUID());
-    setByteFlag(DYNAMICOBJECT_BYTES, 0, type);
+    setByteFlag(DYNAMICOBJECT_BYTES, 0, static_cast<uint8_t>(type));
     setUInt32Value(DYNAMICOBJECT_SPELLID, m_spellProto->getId());
     setFloatValue(DYNAMICOBJECT_RADIUS, radius);
     setUInt32Value(DYNAMICOBJECT_CASTTIME,Util::getMSTime());
