+++ src/world/Units/Creatures/Vehicle.cpp
@@ -175,7 +175,7 @@ void Vehicle::AddPassengerToSeat(Unit* passenger, uint32 seatid)
     {
 #if VERSION_STRING != Cata
         passenger->obj_movement_info.transporter_info.guid = owner->GetGUID();
-        passenger->obj_movement_info.transporter_info.seat = seatid;
+        passenger->obj_movement_info.transporter_info.seat = static_cast<uint8_t>(seatid);
 #endif
     }
 
@@ -588,7 +588,7 @@ void Vehicle::InstallAccessories()
         c->Load(cp, owner->GetPositionX(), owner->GetPositionY(), owner->GetPositionZ(), owner->GetOrientation());
 #if VERSION_STRING != Cata
         c->obj_movement_info.transporter_info.guid = owner->GetGUID();
-        c->obj_movement_info.transporter_info.seat = accessory->seat;
+        c->obj_movement_info.transporter_info.seat = static_cast<uint8_t>(accessory->seat);
 #endif
         c->Phase(PHASE_SET, owner->GetPhase());
         c->SetFaction(owner->GetFaction());
